requirements.txt
fastapi
uvicorn
pydantic
langchain-core
langchain-community
langchain-text-splitters
langchain == 0.0.154
langchain-huggingface
python-dotenv
requests
beautifulsoup4
sentence-transformers
faiss-cpu
streamlit
google.generativeai
playwright
numpy
torch==2.9.0
transformers==4.57.1
huggingface-hub==0.35.3
safetensors
selenium